services/dataset/requirements.txt
Faker>=37.0.0
Pillow>=11.0.0
PyYAML>=6.0.0
reportlab>=4.0.0
